fix bug

+++ ext/mathPro/exercise/game1.php
@@ -14,11 +14,11 @@ function rps($cmd):void
 
     echo 'A player:' . PHP_EOL; //输出字符串 并且换行
     echo '1 mean rock, 2 mean paper, 3 mean scissors' . PHP_EOL;
-    $a = fgets($cmd);//获取A输入的参数
+    $a = (int)fgets($cmd);//获取A输入的参数
 
     echo 'B player:' . PHP_EOL; // 输出字符串并且换行
     echo '1 mean rock, 2 mean paper, 3 mean scissors' . PHP_EOL;
-    $b = fgets($cmd);//获取B输入的参数
+    $b = (int)fgets($cmd);//获取B输入的参数
     $b = -1 * $b; //取反
 
     echo "A is $a and B is $b" . PHP_EOL;
